Move python_check to new router file
