+++ bookserver/routers/coach.py
@@ -1,0 +1,59 @@
+# ***********************************
+# |docname| - Provide code advice
+# ***********************************
+# Endpoints to provide various kinds of advice (syntax/style/etc...)
+# about code samples
+#
+# Imports
+# =======
+# These are listed in the order prescribed by `PEP 8`_.
+#
+# Standard library
+# ----------------
+import ast
+
+# Third-party imports
+# -------------------
+from fastapi import APIRouter, Request
+from pyflakes import checker as pyflakes_checker
+
+# Local application imports
+# -------------------------
+
+# .. _APIRouter config:
+#
+# Routing
+# =======
+# Setup the router object for the endpoints defined in this file.  These will
+# be `connected <included routing>` to the main application in `../main.py`.
+router = APIRouter(
+    # shortcut so we don't have to repeat this part
+    prefix="/coach",
+    tags=["coach"],
+)
+
+
+
+@router.post("/python_check")
+async def python_check(request: Request):
+    """
+    Takes a chunk of Python code and runs a syntax checker (currently
+    Pyflakes) on it to provide more detailed advice than is available
+    via Skulpt.
+    """
+    code_bytes = await request.body()
+    code = code_bytes.decode("utf-8")
+
+    filename = "program.py" 
+
+    resultMessage = ""
+    try:
+        tree = ast.parse(code, filename=filename)
+        w = pyflakes_checker.Checker(tree, filename=filename)
+        w.messages.sort(key=lambda m: m.lineno)
+        for m in w.messages:
+            resultMessage = resultMessage + str(m) + "\n"
+    except SyntaxError as e:
+        resultMessage = filename + ":" + str(e.lineno) + ":" + str(e.offset) + ": " + e.args[0] + "\n"
+
+    return resultMessage
